requirements.txt
PySide6~=6.7.2
future~=1.0.0
pyinstaller~=6.10.0
ffmpeg-python~=0.2.0